refactor seed
so it can handle seed in form of mongo, sqlite, mysql
